relationships

+++ backend/demo/demo/src/main/java/testing/demo/Repository/RequirementsRepo.java
@@ -3,6 +3,6 @@ package testing.demo.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import testing.demo.model.Requirements;
 
-public interface RequirementsRepo extends JpaRepository<Requirements,Integer> {
+public interface RequirementsRepo extends JpaRepository<Requirements,Long> {
     Requirements findById(int id);
 }
